New category groups would be created on every save of the post.
During the process of checking the taxonomy it thinks the given ID is
a string and so it created a new categry group every time a post is
saved.
This is because during the check of the terms the term_id is only
converted to int when the taxonomy is setup as hierarchical.